# src/libs/api-client/model/paginationDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Bookmark Manager App API
 * APIドキュメント
 * OpenAPI spec version: v0.0
 */

export interface PaginationDto {
  total: number;
  count: number;
  offset?: number;
  limit?: number;
}