# src/libs/api-client/model/sampleDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Bookmark Manager App API
 * APIドキュメント
 * OpenAPI spec version: v0.0
 */

export interface SampleDto {
  id: string;
  title: string;
}