# src/libs/api-client/model/updateSampleDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Bookmark Manager App API
 * APIドキュメント
 * OpenAPI spec version: v0.0
 */

export interface UpdateSampleDto {
  /**
   * @minLength 1
   * @maxLength 50
   * @nullable
   */
  title: string | null;
}